// serviceLayer/src/fbs/generated/model/reservationDetailsV2.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */
import type { ILLBibliographicRecord } from "./iLLBibliographicRecord";
import type { Periodical } from "./periodical";

export interface ReservationDetailsV2 {
  dateOfReservation: string;
  /** The date when the patron is no longer interested in the reserved material */
  expiryDate: string;
  ilBibliographicRecord?: ILLBibliographicRecord;
  /** The number in the reservation queue. */
  numberInQueue?: number;
  periodical?: Periodical;
  /** ISIL-number of pickup branch */
  pickupBranch: string;
  /** Set if reserved material is available for loan */
  pickupDeadline?: string;
  /** The reservation number. Will be present if the reservation is ready for pickup (the state is 'readyForPickup') */
  pickupNumber?: string;
  /** The FAUST number */
  recordId: string;
  /** Identifies the reservation for use when updating or deleting the reservation */
  reservationId: number;
  reservationType: string;
  state: string;
  /** Identifies the transaction of reservations */
  transactionId: string;
}